refactor WalletAuth
